Working FFT
