+++ include/matx/core/tie.h
@@ -50,6 +50,22 @@ struct mtie : public BaseOp<mtie<Ts...>>{
 
   static_assert(sizeof...(Ts) > 1, "Only use mtie() when number of elements is greater than 1");
 
+  __MATX_INLINE__ const std::string str() const {
+    if constexpr (sizeof...(Ts) == 2) {
+      return get_type_str(std::get<0>(ts_)) + "=" + get_type_str(std::get<1>(ts_));
+    }
+    else {
+      // Figure this out for proper LHS
+      return "";
+      // std::string lhs = "(";
+      // mapply([&](auto... args) {
+      //   lhs += ((get_type_str(args) + ","), ...);
+      // }, ts_);
+
+      // return lhs.substr(0, lhs.size() - 1) + ")";
+    }
+  }  
+
   mtie(Ts... ts) : ts_(ts...) {}
 
   // operator= to cover multiple return types using mtie in a run statement
@@ -92,59 +108,5 @@ struct mtie : public BaseOp<mtie<Ts...>>{
   std::tuple<Ts...> ts_;
 };
 
-// /* mtie exists because we can't overload std::tuple's operator= for operator assignments in
-//    a multiple return set statement. */
-// template <typename TieType, typename XformOp>
-// struct mtie_container {
-//   using mtie_type = bool;
-//   using scalar_type = void; // Doesn't matter since it's not used
-//   using value_type  = void; // Doesn't matter since it's not used
-//   using shape_type  = index_t;
-//   using matxoplvalue = bool;
-
-//   mtie_container(TieType &&t, XformOp &&xop) : tie_(t), xop_(xop) {}
-
-
-//   template <int n>
-//   auto get() {
-//     return std::get<n>(ts_);
-//   }
-
-//   static __MATX_INLINE__ constexpr int32_t Rank()
-//   {
-//     return 1;
-//   }
-
-//   constexpr __MATX_INLINE__ __MATX_HOST__ __MATX_DEVICE__ auto Size([[maybe_unused]] int dim) const noexcept
-//   {
-//     return 0;
-//   }  
-
-//   constexpr __MATX_INLINE__ auto Elements() const noexcept {
-//     return std::tuple_size(ts_);
-//   }
-
-//   constexpr __MATX_INLINE__ auto IsSingleAssign() const noexcept { 
-//     return sizeof...(Ts) == 2;
-//   }
-
-//   template <typename Executor>
-//   __MATX_INLINE__ void Exec(Executor &&ex) {
-//     if constexpr (sizeof...(Ts) == 2) {
-//       std::get<1>(ts_).Exec(std::get<0>(ts_), std::forward<Executor>(ex));
-//     }
-//     else {
-//       // When we're not doing a direct assignment (ie (A = xform(B)) the last element of
-//       // our tuple is the transform operator and the first N elements are the output parameters
-//       // for the transform.
-//       mapply([&](auto... args) {
-//         std::get<sizeof...(Ts) - 1>(ts_).Exec(args..., std::forward<Executor>(ex));
-//       }, ts_);
-//     }
-//   }
-
-//   TieType tie_;
-//   XformOp xop_;
-// };
 
 }; // namespace matx
